### 2024-03-16 #6497 @osorensen
> ```
> Reference check summary (note 'MISSING' DOIs are suggestions that need verification):
>
> OK DOIs
>
> - 10.21105/joss.03434 is OK
> - 10.21105/joss.03326 is OK
> - 10.1109/MSP.2018.2869928 is OK
> - 10.1109/MSP.2018.2875349 is OK
> - 10.1145/2964284.2973795 is OK
> - 10.1007/978-3-319-21945-5 is OK
> - 10.25080/Majora-7b98e3ed-003 is OK
> - 10.5281/zenodo.1415016 is OK
> - 10.5281/zenodo.1416032 is OK
> - 10.5281/zenodo.849741 is OK
> - 10.1145/1631272.1631459 is OK
> - 10.5281/zenodo.1417145 is OK
>
> MISSING DOIs
>
> - No DOI given, and none found for title: libf0: A Python Library for Fundamental Frequency ...
> - No DOI given, and none found for title: Tempogram Toolbox: MATLAB Tempo and Pulse Analysis...
> - No DOI given, and none found for title: Sonification Report: Status of the Field and Resea...
>
> INVALID DOIs
>
> - None
> ```

@yiitozer can you please fix the missing DOIs and notify me here when done?
