### 2024-03-17 4f091c4
Update paper.md
There seemed to be a double space between "an" and "audio", causing a lineshift on the compiled pdf. Trying to fix this here.
