### 2024-03-18 35f2db9
Update references.bib
add doi
