+++ paper/references.bib
@@ -4,6 +4,7 @@
   booktitle = {Demos and Late Breaking News of the International Society for Music Information Retrieval Conference ({ISMIR})},
   address   = {Bengaluru, India},
   year      = {2022},
+  doi       = {10.5281/zenodo.7512227}
 }
 
 @article{MuellerOKPD21_SyncToolbox_JOSS,
