### 2024-03-18 #6497 @yiitozer
Dear @osorensen, I fixed one of the DOIs (for the f0 toolbox), however I couldn't find one for the other two...

35f2db969a6ca47a227835a9aaf9e957749cdb95
