### 2024-03-18 #6497 @osorensen
> Dear @osorensen, I fixed one of the DOIs (for the f0 toolbox), however I couldn't find one for the other two... [groupmm/libsoni@35f2db9](https://github.com/groupmm/libsoni/commit/35f2db969a6ca47a227835a9aaf9e957749cdb95)

That seems fine. Thanks!
