### 2024-03-22 #6524 @editorialbot
[REVIEW]: libsoni: A Python Toolbox for Sonifying Music Annotations and Feature Representations
**Submitting author:** @yiitozer (<a href="http://orcid.org/0000-0003-2235-8655">Yigitcan Özer</a>)
**Repository:** https://github.com/groupmm/libsoni
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.0.2
