### 2024-04-16 #6524 @ren-zeng
- The authors mentioned the following performance statement. However, there is no empirical support for improved efficiency.
  - > Secondly, we re-implemented and included all necessary components (e.g., sound generators based on sinusoidal models and
click sounds), even though similar basic functionality is available in librosa and libfmp. By doing so, libsoni offers a coherent API along with convenient but easily modifiable parameter
presets. Additionally, the implementations are more efficient than previous software.

  - The API is satisfactory, it would be better if
      - there are minimal examples for each core function in the documentation.
      - all core functions specify the expected shape of the input numpy array (and what each dimension means) and the shape of the resulting array.
- [ ] Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?
  - There are tests but they cover very limited scenarios of the input data. It would be better if
    - the authors use a more comprehensive testing framework like unittest and use randomly generated input instead of hand-coded minimal examples.
    - the test input arrays are non-trivial and have more complex dimensions.
    - the test covers more negative examples so that the user can know what properties the inputs need to satisfy.
