### 2024-04-19 #6524 @yiitozer
Dear @ren-zeng, thank you very much for your constructive suggestions.

1. Performance:
We will include a comparative analysis with librosa and libfmp in the Jupyter Notebooks.

2. Functionality documentation:
* We indeed provided minimal examples per core function in the Jupyter notebooks. Shall we also integrate them to the API?
* The explicit explanation of the dimension is a good point, we will add the shapes of the resulting arrays.

3. Automated tests:
* Since the toolbox is designed to sonify input annotations, we chose not to use randomly generated inputs to have a controlled scenario. Could you please elaborate on this suggestion?
* We will also add tests for the spectrogram case, where the input dimensions are more complex.
* We will add negative examples, thank you for your suggestion.
