### 2024-04-24 7fa52dd
bugfix: sonification_duration (should be in samples)
