+++ libsoni/core/pianoroll.py
@@ -155,7 +155,7 @@ def sonify_pianoroll_clicks(pianoroll_df: pd.DataFrame,
 def sonify_pianoroll_sample(pianoroll_df: pd.DataFrame,
                             sample: np.ndarray = None,
                             reference_pitch: int = 69,
-                            sonification_duration: float = None,
+                            sonification_duration: int = None,
                             signal_fading_duration: float = 0.05,
                             note_fading_duration: float = 0.01,
                             normalize: bool = True,
@@ -176,8 +176,8 @@ def sonify_pianoroll_sample(pianoroll_df: pd.DataFrame,
     reference_pitch: int, default = 69
         Original pitch of the sample.
 
-    sonification_duration: float, default = None
-        Determines duration of sonification, in seconds.
+    sonification_duration: int, default = None
+        Determines duration of sonification, in samples.
 
     signal_fading_duration: float, default = 0.05
         Determines duration of fade-in and fade-out at beginning and end of the final sonification, in seconds.
@@ -226,7 +226,7 @@ def sonify_pianoroll_fm_synthesis(pianoroll_df: pd.DataFrame,
                                   mod_rate_relative: float = 0.0,
                                   mod_amp: float = 0.0,
                                   tuning_frequency: float = 440.0,
-                                  sonification_duration: float = None,
+                                  sonification_duration: int = None,
                                   signal_fading_duration: float = 0.05,
                                   note_fading_duration: float = 0.01,
                                   normalize: bool = True,
@@ -251,8 +251,8 @@ def sonify_pianoroll_fm_synthesis(pianoroll_df: pd.DataFrame,
     tuning_frequency: float, default = 440.0
         Tuning frequency in Hertz.
 
-    sonification_duration: float, default = None
-        Determines duration of sonification, in seconds.
+    sonification_duration: int, default = None
+        Determines duration of sonification, in samples.
 
     signal_fading_duration: float, default = 0.05
         Determines duration of fade-in and fade-out at beginning and end of the final sonification, in seconds.
@@ -298,23 +298,23 @@ def sonify_pianoroll_fm_synthesis(pianoroll_df: pd.DataFrame,
 
 def __init_pianoroll_sonification(pianoroll_df: pd.DataFrame,
                                   fs: int,
-                                  sonification_duration: float = None):
+                                  sonification_duration: int = None):
 
     pianoroll_df = format_df(pianoroll_df)
     num_samples = int(pianoroll_df['end'].max() * fs)
 
     if sonification_duration is None:
-        sonification_duration = pianoroll_df['end'].max()
-    sonification_duration_samples = int(sonification_duration * fs)
+        sonification_duration = int(pianoroll_df['end'].max()) * fs
+
+    sonification_duration_secs = sonification_duration / fs
 
     # if sonification_duration is less than num_samples, crop the arrays
-    if sonification_duration_samples < num_samples:
-        pianoroll_df = pianoroll_df[pianoroll_df['start'] < sonification_duration]
-        pianoroll_df.loc[pianoroll_df['end'] > sonification_duration, 'end'] = sonification_duration
+    if sonification_duration < num_samples:
+        pianoroll_df = pianoroll_df[pianoroll_df['start'] < sonification_duration_secs]
+        pianoroll_df.loc[pianoroll_df['end'] > sonification_duration_secs, 'end'] = sonification_duration_secs
 
     pianoroll_df['duration'] = pianoroll_df['end'] - pianoroll_df['start']
 
-    num_samples = sonification_duration_samples
-    pianoroll_sonification = np.zeros(num_samples)
+    pianoroll_sonification = np.zeros(sonification_duration)
 
     return pianoroll_df, pianoroll_sonification
