### 2024-04-24 0643207
bugfix: sonification_duration (should be in samples)
