### 2024-04-24 ea3fdd4
nbconvert is not a mandatory package...
