### 2024-04-24 45d5b04
unit test update: convert to unit tests, add negative examples.
