### 2024-04-24 1a135d0
spectrogram tests added.
