### 2024-04-24 abefb9c
nbextensions not needed.
