+++ environment_libsoni.yml
@@ -16,7 +16,6 @@ dependencies:
   - scikit-learn >=0.21.0, <1.0.0
   - ffmpeg >=2.8.0, <3.0.0
   - numba >=0.51.0, <1.0.0
-  - jupyter_contrib_nbextensions >=0.5.0, <1.0.0
   - pip:
     - mir-eval >=0.5.0, <1.0.0
     - nbstripout >=0.3.0, <1.0.0
