### 2024-04-24 #6524 @yiitozer
Dear @ren-zeng,

Performance:
We have chosen to remove this statement, as the basic functions provided by libfmp and librosa show similar performances as those by libsoni:
https://github.com/groupmm/libsoni/commit/859ea29843cd7a886f098e763baa8ef5e60c9458

Functionality documentation:
We updated the API documentation indicating the explicit shapes of input and output arrays. Please see:

Since we provided minimal examples of the core functions in the notebooks, adding those to the API documentation
