### 2024-04-28 #6524 @ren-zeng
Dear @yiitozer,

Thank you for your update on the package and the paper. On my side, everything looks good now. Regarding my comment on the randomly generated inputs, I was mainly concerned, as a potential user, about what kinds of inputs are valid. But since you already added the required array shapes, it is not necessary anymore.

Hi @osorensen, I have checked all the boxes and my review is finished.
