### 2024-04-28 #6524 @expectopatronum
I am deeply sorry for taking so long! I (once again) overestimated how much time I would have with a toddler at home 🙃

After solving a few issues directly by opening a Github issue, there are only a few things left:
- Documentation: It took me a while to find the [link to the documentation](https://groupmm.github.io/libsoni/build/html/index.html) (in the sidebar of the github repo), maybe it could also be linked somewhere directly in the README?
- For me it is not entirely clear whether by "Documentation" (in the review check list) the linked documentation is meant or the README + the linked documentation. In the first case I cannot check some of the boxes, because the installation instructions, example usage, ... is not described in what I think is the "documentation". Maybe it would be sufficient to mention in the starting page of the documentation that more details (installation instructions, ...) can be found in the Github repository. I'm thinking of the case where someone finds the package github.io page and does not know about the repository.
- I did have the same question about the performance claims as @ren-zeng but this is solved now.
- This is a very minor thing and I was not even sure if I should mention it but it bugs me a bit that both "Licence" and "License" are used.

Besides that, the package is great and I would have loved to have something like this when I started my PhD! I think it will be very useful for research in MIR and I think it will be easy to get started with it thanks to the example notebooks (which is not surprising coming from this group :)).

Thanks for your patience and best regards
Verena
