+++ README.md
@@ -71,8 +71,8 @@ directly contacting us via email (meinard.mueller@audiolabs-erlangen.de) or for 
 an issue in our GitHub repository. Please do not submit a pull request without prior consultation
 with us.
 
-## Licence
-The code for this toolbox is published under an [MIT licence](LICENCE).
+## License
+The code for this toolbox is published under an [MIT license](LICENSE).
 This does not apply to the data files:
 * Schubert songs are taken from the [Schubert Winterreise Dataset](https://zenodo.org/record/4122060). 
 * Recording of the cantata *Ach Gott und Herr* by Bach is taken fom [Bach10 Dataset](https://labsites.rochester.edu/air/datasets/Bach10%20Dataset_v1.0.pdf).
