### 2024-04-29 #6524 @yiitozer
Dear @expectopatronum,

>     * Documentation: It took me a while to find the [link to the documentation](https://groupmm.github.io/libsoni/build/html/index.html) (in the sidebar of the github repo), maybe it could also be linked somewhere directly in the README?
>
>     * For me it is not entirely clear whether by "Documentation" (in the review check list) the linked documentation is meant or the README + the linked documentation. In the first case I cannot check some of the boxes, because the installation instructions, example usage, ... is not described in what I think is the "documentation". Maybe it would be sufficient to mention in the starting page of the documentation that more details (installation instructions, ...) can be found in the Github repository. I'm thinking of the case where someone finds the package github.io page and does not know about the repository.

Thank you very much for your suggestions, we added a link to the documentation in the README file and a link to the GitHub repository in the API documentation, as well as renamed licence to license :) : https://github.com/groupmm/libsoni/pull/13

> Besides that, the package is great and I would have loved to have something like this when I started my PhD! I think it will be very useful for research in MIR and I think it will be easy to get started with it thanks to the example notebooks (which is not surprising coming from this group :)).
>
> Thanks for your patience and best regards Verena

Thank you very much for your constructive and motivating comment!! :) We're very happy to get such feedback!

Best,
Yigit
