### 2024-04-29 #6524 @yiitozer
Dear @osorensen, thank you very much for the detailed instructions.

* We have tagged the release of the software as v1.0.2 and published it on PyPI: https://pypi.org/project/libsoni/

* Archived the reviewed software on Zenodo: https://zenodo.org/records/11085871

* The DOI of the archived version: 10.5281/zenodo.11085871
