### 2024-04-29 #6524 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
