### 2024-05-03 #6524 @yiitozer
Dear @osorensen, is the paper than acceptable with the command "editorialbot accept"? :)
