### 2024-05-05 #6524 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
