### 2024-05-05 b723ce0
Update paper.md
Minor fix for parenthetical citation.
