### 2024-05-05 #6524 @oliviaguest
@yiitozer please review the PR (https://github.com/groupmm/libsoni/pull/14) and then recompile the paper and let me know i this is a version you are happy to publish. ☺️
