### 2024-05-07 #6524 @oliviaguest
I am really confused by this @arfon, how can this be merged here: https://github.com/groupmm/libsoni/commit/48297aa2cc986be5107a325c2ef87c6b3dd62e58 and then the PDF contains the double brackets nonetheless (see lines 132-133 in PDF: https://github.com/openjournals/joss-papers/blob/joss.06524/joss.06524/10.21105.joss.06524.pdf)? What are we missing?
