### 2024-05-08 #6524 @arfon
@oliviaguest was something weird being cached here? Does this look correct to you?

![Image](https://github.com/openjournals/joss-reviews/assets/4483/d69b848e-25a2-4b81-9b3c-152901a8df37)
