### 2024-05-26 #6524 @oliviaguest
@mstimberg said:

> Adding this to the documentation has been on my todo list for a long time: I think the correct syntax here is [e.g., using alignment methods provided by the Sync Toolbox, @MuellerOKPD21_SyncToolbox_JOSS] which renders as "(e.g., using alignment methods provided by the Sync Toolbox, Müller et al., 2021)".
